refactor: isolate pending write buffer helper
